+++ sql_tools/mysql/__init__.py
@@ -3,7 +3,7 @@ from .execute import execute
 from .dbTools import *
 from .tableTools import *
 
-__version__ = f"SQL-Tools: MySQL version: 0.1"
+__version__ = f"SQL-Tools: MySQL version: 1.1"
 __help__ = 'Visit the documentation for more help or type "help(sql_tools)"'